--- mock/user/[id].ts
@@ -1,19 +1,0 @@
-import { MockMethod } from '../../src';
-
-export const get: MockMethod = (req, res) => {
-	const { id } = req.query || {};
-	res.end(
-		JSON.stringify({
-			id,
-			name: 'qingyang11',
-		}),
-	);
-};
-
-export const post: MockMethod = (req) => {
-	const { id } = req.query || {};
-	return {
-		id,
-		name: 'qingyang',
-	};
-};
