+++ package.json
@@ -10,7 +10,7 @@
   ],
   "type": "module",
   "engines": {
-    "node": "^14.20.0"
+    "node": "^14.19.0"
   },
   "keywords": [
     "vite",
@@ -31,29 +31,22 @@
   },
   "homepage": "https://github.com/minjs1cn/vite-plugin-best-mock/tree/main/#readme",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc",
+    "dev": "pnpm build && pnpm -F playground dev",
+    "build": "pnpm -r build",
     "preview": "vite build",
     "release": "standard-version"
   },
   "peerDependencies": {
-    "mockjs": ">=1.1.0",
-    "vite": ">=2.0.0"
+    "mockjs": ">=1.1.0"
   },
   "dependencies": {
-    "chalk": "^5.1.2",
-    "colors": "^1.4.0",
-    "connect-multiparty": "^2.2.0",
-    "qs": "^6.11.0",
-    "ts-node": "^10.9.1",
     "typescript": "^4.8.4",
     "vite": "^3.1.8"
   },
   "devDependencies": {
     "@types/multiparty": "^0.0.33",
     "@types/node": "^18.11.0",
-    "@types/qs": "^6.9.7",
-    "pkg-get": "^0.0.2",
+    "rimraf": "^3.0.2",
     "standard-version": "^9.5.0"
   }
 }
