+++ packages/core/src/connect-multiparty.d.ts
