+++ packages/core/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "outDir": "lib",
+    "esModuleInterop": true
+  }
+}
