+++ packages/playground/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Vite Plugin Best Mock</title>
+</head>
+
+<body>
+  <input type="file" id="file" />
+  <script type="module" src="./lib/main.js"></script>
+</body>
+
+</html>
