+++ packages/playground/mock/user/index.ts
@@ -1,0 +1,17 @@
+import { MockMethod } from 'core';
+
+export const get: MockMethod = (req) => {
+  const { id } = req.query || {};
+  return {
+    id,
+    name: '清扬',
+  };
+};
+
+export const post: MockMethod = (req) => {
+  const { id } = req.body || {};
+  return {
+    id,
+    name: '清扬',
+  };
+};
