--- playground/a.ts
@@ -1,1 +1,0 @@
-alert('1');
